+++ magic_admin/resources/__init__.py
@@ -1,3 +1,4 @@
 from magic_admin.resources.token import Token  # noqa: F401
 from magic_admin.resources.user import User  # noqa: F401
 from magic_admin.resources.wallet import WalletType  # noqa: F401
+from magic_admin.resources.utils import Utils  # noqa: F401